libnetlink: Make sure there is no error happened when dumping is done
If the kernel is misconfigured (say UDP diag module is not set)
we should detect such situation and report an error.

This patch adds a test for error code. Note the kernel requires
a patch as well. Without kernel patch this test do not harm
anyway, thus safe to merge.

#2412

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>