crtools: rename binary to criu
"Tools" in the context of a binary name doesn't make any sense,
because every binary is a tool (lstool, dftool), plus it's one
tool.

So let's rename the beast to criu, which is short, pretty unique
and coincides with the name of the project.

Backward-compatible symlink is added for the sake of tests,
we can fix them later and remove the link.

Packagers beware: our binary is no longer crtools but criu.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>