criu: Version 0.5 release
This is lots-of-bug-fixes release -- many new people started playing
with criu and reported a lot of issues with it.

Other than this, we have several new features. First of all it's C/R
of netlink sockets C/R (required for Fedora container dump) and pending
signals.

Next, we've changed the format of memory dumps. This allowed us to
implement so called page server, that allows to send memory dumps
directly on another box w/o intermediate dump file. Also this new
format made it possible to implement the (not yet final) --snapshot
option, that demonstrates the work of user memory changes tracker
in the kernel.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>