criu --help: do not use tabs
Some lines were using tabs, some were not. Personally I am not in favor
of any style, but let's be consistent in either using tabs or not.

Changing tabs with spaces produced a smaller patch, so let's do this
way.

This patch should not produce any visual change in output.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>