arm: reset the CPSR.T bit when executing the infected code
The infected code must be executed in the ARM mode but the dumpee
may run in the Thumb mode while being dumped so the bit CPSR.T
may be set when the control is transfered to the infected blobs
so this code is executed in the Thumb mode too.

We have to clear the bit CPSR.T before transfering control
to the infected code to prevent the behavior described above.

Signed-off-by: Alexander Kartashov <alekskartashov@parallels.com>
Reported-by: Chanho Park <chanho61.park@samsusng.com>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>