parasite: Restore parasite state via rt_sigreturn
This patch reduces a window, when a crtools can kill a dumped process,
because if a parasite in a deamon mode can restore the state of the
process, if crtools detached unexpectedly.

All threads are synchronized on the _exit_ point from sys_rt_sigreturn,
for that crtools traces all syscalls. After that we remove the parasite
blob from dumpee as we did it previously and let the process run further.

Signed-off-by: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>